backend/requirements.txt
Flask
flask-cors
psycopg2-binary
Pillow
google-generativeai
google-genai
google-search-results
celery
redis
requests
python-dotenv
gevent
